backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
